+++ test/test_helper.rb
@@ -26,6 +26,7 @@ require "minitest/autorun"
 
 # GLOBAL TOKEN VARIABLES
 # Configure valid tokens so that the tests can be carried out successfully.
-@@meli_access_token = 'APP_USR-7241495617113935-012117-2ad6a02378dbdc38d235e308323d5c20-1632856741'
-@@meli_refresh_token = 'TG-65ad893589af6800018e6885-1632856741'
 @@meli_user_id = '1632856741'
+@@meli_item_id = 'MLB4389178328'
+@@meli_access_token = 'APP_USR-7241495617113935-012210-29222f14fe4b96067fd8e34c571e3dbf-1632856741'
+@@meli_refresh_token = 'TG-65ae76825114eb0001e1cd24-1632856741'